requirements.txt
# Core ML / DL
tensorflow==2.15.0
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2

# Image processing
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# Visualisation (optional at runtime, needed for notebook)
matplotlib==3.8.4
seaborn==0.13.2

# Utilities
eli5==0.13.0
tqdm==4.66.4

# Testing
pytest==8.2.0
pytest-cov==5.0.0
